requirements.txt
boto3==1.23.9
botocore==1.26.10
cffi==1.15.0
click==8.0.3
cryptography==36.0.1
Flask==2.0.2
Flask-Cors==3.0.10
flask-marshmallow==0.14.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
itsdangerous==2.0.1
Jinja2==3.0.3
jmespath==1.0.0
MarkupSafe==2.0.1
marshmallow==3.14.1
marshmallow-sqlalchemy==0.26.1
passlib==1.7.4
pycparser==2.21
PyJWT==1.7.1
python-dateutil==2.8.2
python-dotenv==0.20.0
s3transfer==0.5.2
six==1.16.0
SQLAlchemy==1.4.28
urllib3==1.26.9
uuid==1.30
Werkzeug==2.0.2
